feat: Add nanosecond timestamps, UTF-8 cache, and fix dependency direction
Performance optimizations and critical architectural fix:

- Nanoseconds branded type: BigInt-based for true nanosecond precision
  in Node.js via process.hrtime.bigint()
- UTF-8 cache (SIEVE algorithm): Efficient string->Uint8Array caching
  for Arrow conversion with LRU-like eviction
- Thread ID: Unique thread/worker identification for distributed tracing
- Arrow conversion rewrite: Complete overhaul for zero-copy conversion
  with proper dictionary encoding for enum/category types
- CRITICAL: Fixed dependency direction - arrow-builder MUST NOT import
  from lmao (was incorrectly allowing bidirectional imports)
- Spec updates: Enhanced buffer architecture, self-tuning, and
  tree walker documentation

+++ packages/lmao/src/lib/__tests__/arrow-builder-integration/buffer-chaining.test.ts
@@ -94,7 +94,7 @@ describe('Buffer Chaining', () => {
       expect(nextBuffer['attr_duration']).toBeInstanceOf(Float64Array);
 
       // Should have core columns
-      expect(nextBuffer.timestamps).toBeInstanceOf(Float64Array);
+      expect(nextBuffer.timestamps).toBeInstanceOf(BigInt64Array);
       expect(nextBuffer.operations).toBeInstanceOf(Uint8Array);
     });
